Move leftover player logic from board to player
