+++ src/main/java/inf112/roborally/game/board/TiledBoard.java
@@ -1,12 +1,13 @@
 package inf112.roborally.game.board;
 
+import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TiledMapRenderer;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
 
-public abstract class BoardCreator {
+public abstract class TiledBoard {
 
     protected TiledMap map;
     protected TmxMapLoader loader;
@@ -27,6 +28,11 @@ public abstract class BoardCreator {
         createLayers();
     }
 
+    public void render(OrthographicCamera camera) {
+        mapRenderer.setView(camera);
+        mapRenderer.render();
+    }
+
     protected void createLayers() {
         beltLayer = (TiledMapTileLayer) map.getLayers().get("belts");
         floorLayer = (TiledMapTileLayer) map.getLayers().get("floor");
@@ -35,5 +41,37 @@ public abstract class BoardCreator {
         startLayer = (TiledMapTileLayer) map.getLayers().get("start");
     }
 
+    public void dispose() {
+        System.out.println("disposing board");
+        map.dispose();
+
+    }
+
+    public TiledMapTileLayer getLaserLayer() {
+        return laserLayer;
+    }
+
+    public TiledMapTileLayer getStartLayer() {
+        return startLayer;
+    }
+
+    public TiledMapTileLayer getBeltLayer() {
+        return beltLayer;
+    }
+
+    public TiledMapTileLayer getWallLayer() {
+        return this.wallLayer;
+    }
+    public TiledMapTileLayer getFloorLayer() {
+        return floorLayer;
+    }
+
+    public int getWidth() {
+        return this.floorLayer.getWidth();
+    }
+
+    public int getHeight() {
+        return this.floorLayer.getHeight();
+    }
 
 }
